Add support for Ec2Snitch.
The plumbing to add Ec2Snitch to cassandra container, the entrypoint for
the image already supports passing in any form of snitch into
`/etc/cassandra/cassandra.yaml, the dc_suffix is optional based on the
code below:

https://github.com/apache/cassandra/blob/a85afbc7a83709da8d96d92fc4154675794ca7fb/src/java/org/apache/cassandra/locator/Ec2Snitch.java#L65

and the docs at:

https://docs.datastax.com/en/archived/cassandra/3.0/cassandra/architecture/archSnitchEC2.html

So we do not need extra plumbing in the entrypoint script for the image.

This also re-implements the hack from Ec2Snitch:

https://github.com/apache/cassandra/blob/a85afbc7a83709da8d96d92fc4154675794ca7fb/src/java/org/apache/cassandra/locator/Ec2Snitch.java#L61-L63

so we can pass the correct datacenter to the backend's config.