DEVOPS-1805: Build configurator images from Jenkins.
This will build configurator images from Jenkins that includes all of
our configuration, which will eliminate the need for the user to clone
and build the images themselves.

Currently I plan to make this a single branch Jenkins pipeline and build
only images from a "blessed" branch, if we find need for building all
branches we can revisit this decision.