DEVOPS-1805 Simply steps to run configurator.
This relies on only the presence of a values.yaml in current working
directory and leverages the configs add to the image to generate
manifests. This technically cuts away 2 steps from the steps required to
run the configurator, since there is no need for the user to clone repo
and build the image themselves.

Anticipated usage:

`docker run -v ${PWD}:/manifests --rm -it ${img}`

Depends on:

https://github.com/draios/sysdigcloud-kubernetes/pull/194