Fix helm templates and values.yaml.
`default` function in helm only works when the top level key is absent,
for hierarchical data, the hierarchy must be present and `default` will
fill in if the value is absent.

This also addressed lookups that had missing defaults.

This was an oversight I did not test after commenting out the fields, as
it currently is the fields cannot be commented out otherwise things will
break.