DEVOPS-1805 Support local path via k8s hostPath volumes.
This does a couple of things:

- Introduces one more option to the union field for
storageClassProvisioner `hostPath`. I avoided naming this `local`
because we might end up needing to support the
[`local`](https://kubernetes.io/docs/concepts/storage/storage-classes/#local).
Caveat: hostPath only works with small cluster types.
provisioner, currently this mode is only enabled for small clusters.
- cleans up verifying the existence of the `STORAGE_CLASS` in
generate_templates.sh, this was failing when the storage class does not
exist and because we set the `-e`(errexit) flag this causes the script
to fail when the storage class does not exist, the workaround is to pipe
stderr to `/dev/null` and or the command to `/bin/true` preventing the
early exit.
- Moved volumeClaimTemplates towards the bottom in a couple of places to
make the conditional easier to write.