Introduce sane defaults for most of values in values.yaml (#277)
* Make the default action for configurator to be generate+deploy.

Previously configurator did not have a default action to run when a user
invoked it, ideally when a user invokes it they either want to generate
yaml, deploy yaml or generate and deploy yaml, given this limited set of
options it seems easy to provide a default for users.

This also changes the way "generate and deploy" is configured from
`generate deploy` to `generate+deploy` which makes it look more like
what it is. It also makes sure `readConfigFromValuesYaml` includes
`defaultValues.yaml` when looking up values.