Create service of type LoadBalancer only when explicitly configured. (#292)
Previously we always created a service of type LoadBalancer, usually
this NodePort behind the LoadBalancer is only what is used and not the
LoadBalancer itself, it also causes error events that are confusing to
the user. This makes creation of the LoadBalancer configurable and the
user can explicitly request for a LoadBalancer if they want one,
otherwise we create a NodePort by default.