Make hostPath host directories configurable. (#295)
This makes hostPath host directories configurable and makes it available
to cluster sizes larger than small. It also makes the hostname based
podAntiAffinity requiredDuringSchedulingIgnoredDuringExecution for
hostPath based storageClassProvisioner, this ensures pods of the same
replica do not end up on the same host and start clobbering their data.

See configuration.md changeset for coverage on how the new feature is
configured.

Concerns (which is not entirely new since folks are already using
hostPath in production), any event that causes deletion of a pod will
require the admin to ensure the pod is replaced on the node it
previously was otherwise there will be chaos.